ch12/requirements.txt
certifi==2018.4.16
chardet==3.0.4
confusable-homoglyphs==3.1.1
defusedxml==0.5.0
Django==1.8.13
django-allauth==0.26.1
django-cart==1.0.4
django-email-confirmation==0.2
django-filer==1.3.2
django-mailgun==0.9.1
django-markdown-deux==1.0.5
django-mptt==0.8.7
django-polymorphic==1.3.1
django-ranged-response==0.2.0
django-registration==2.4.1
django-simple-captcha==0.5.6
django-uni-form==0.9.0
easy-thumbnails==2.5
httplib2==0.11.3
idna==2.6
markdown2==2.3.5
mysqlclient==1.3.12
oauth2==1.9.0.post1
oauthlib==2.1.0
Pillow==5.1.0
pkg-resources==0.0.0
python-openid==2.2.5
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
Unidecode==0.4.21
urllib3==1.22
